refactor(utxo-lib): composeTx merge `coinselect` modules
